search_score()実行時に固まるバグをThreadingを用いて修正
