+++ .gitignore
@@ -23,3 +23,4 @@ error_indent.mjson
 d2.mjson
 success_indent.mjson
 >>>>>>> develop
+mobiDB_human.mjson
