+++ screen_wait_search.py
@@ -2,7 +2,8 @@ from logging import getLogger, StreamHandler, DEBUG
 import time
 import json
 import config
-
+import threading
+from kivy.app import App
 
 """デバック"""
 logger = getLogger(__name__)
@@ -13,13 +14,13 @@ logger.addHandler(handler)
 logger.propagate = False
 
 
-class SearchScore:
-    def search_score(self):
+class SearchScore(threading.Thread):
+    def run(self):
         logger.debug("screen_wait_search, SearchScore, search_score()")
-        
+
         # jsonファイル読み込み，条件比較を行う
         with open('success_data.mjson', 'w') as fw:
-            with open("disorder_add_protain.mjson", "r") as fr:
+            with open("mobiDB_human.mjson", "r") as fr:
                 t1 = time.time()
 
                 for (i, line) in enumerate(fr):
@@ -29,11 +30,33 @@ class SearchScore:
                     ignored_times = 0
                     pos = config.threshold_len
 
-                    try:
-                        # keywordが含まれているか判定
-                        if config.keyword not in json_dict["protein names"]:
-                            continue
+                    """
+                            score[pos]              : pos番目のscore値。
+                            config.threshold_val    : score値用の閾値。
+                            config.threshold_len    : 閾値以上が何回続けばよいかを決める変数。
+                            succeeded_times         : scores[pos] > config.threshold_val が True であった回数を保持する変数。
 
+                            whileループは　succeeded_times > threshold_len のときに抜ける。
+
+                            config.fill_gap         : 閾値以下を何回まで許すかを決める変数　
+                                                      例）　score[ 1, 1, 0, 0, 1, 1 ], threshold_val = 0.5 とする。
+                                                         ---------------------------------------------
+                                                            for i in len(score):
+                                                                if score[i] > threshold_val:
+                                                                    i += 1
+                                                                elif score[i] < threshold_val
+                                                                    i = 0
+                                                         ---------------------------------------------
+                                                         上記のように比較した場合の出力は、
+                                                            fill_gap == 1 -> 2
+                                                            fill_gap == 2 -> 6
+                                                         と違いがでる。
+                                                         このように、閾値以下を何回許すかを決める変数をfill_gapとする。
+
+                            ignored_times           : 無視した回数の合計を保持する変数。
+                            """
+
+                    try:
                         scores = json_dict["mobidb_consensus"]["disorder"]["predictors"][1]["scores"]
 
                         while pos < len(scores):
@@ -61,32 +84,14 @@ class SearchScore:
                     except IndexError:
                         pass
 
+                self.change_screen("out")
+
                 t2 = time.time()
                 elapsed_time = t2 - t1  # 処理にかかった時間を計算する
                 print("経過時間：", elapsed_time)
 
-        """
-        score[pos]              : pos番目のscore値。
-        config.threshold_val    : score値用の閾値。
-        config.threshold_len    : 閾値以上が何回続けばよいかを決める変数。
-        succeeded_times         : scores[pos] > config.threshold_val が True であった回数を保持する変数。
+    def change_screen(self, name):
+        logger.debug("screen_wait_main.py, ScreenWait, change_screen()")
 
-        whileループは　succeeded_times > threshold_len のときに抜ける。
-
-        config.fill_gap         : 閾値以下を何回まで許すかを決める変数　
-                                  例）　score[ 1, 1, 0, 0, 1, 1 ], threshold_val = 0.5 とする。
-                                     ---------------------------------------------
-                                        for i in len(score):
-                                            if score[i] > threshold_val:
-                                                i += 1
-                                            elif score[i] < threshold_val
-                                                i = 0
-                                     ---------------------------------------------
-                                     上記のように比較した場合の出力は、
-                                        fill_gap == 1 -> 2
-                                        fill_gap == 2 -> 6
-                                     と違いがでる。
-                                     このように、閾値以下を何回許すかを決める変数をfill_gapとする。
-
-        ignored_times           : 無視した回数の合計を保持する変数。
-        """
+        app = App.get_running_app()
+        app.sm.current = name
